+++ tests/testthat/helper-requiet.R
@@ -1,4 +1,5 @@
 requiet <- function(package) {
+  testthat::skip_if_not_installed(package)
   suppressPackageStartupMessages(
     require(package, warn.conflicts = FALSE, character.only = TRUE)
   )