+++ NEWS.md
@@ -30,8 +30,10 @@
   
 ## Bug Fixes
 
-* Fixes wrong *z*-score values when using a vector instead of a data frame in
+* Fixed wrong *z*-score values when using a vector instead of a data frame in
   `check_outliers()` (#476).
+
+* Fixed `cronbachs_alpha()` for objects from `parameters::principal_component()`.
 
 # performance 0.9.2